Move OSystem::Random -> System::Random for determinism + state deprecations
The ALE is based on v2.4.2 of Stella. Shortly after the 2.4.2 release, the Stella authors moved the RNG from OSystem -> System. This was done to enable cartridges to randomize entries on **reset** instead of construction. This more accurately emulates the hardware and for our purposes allows us to enable full determinism. Currently, we have two sources of emulator randomness

1) Cartridge RAM randomization
2) RAM I/O Timer

Currently, in the ALE, we were using the same RNG for emulator and environment randomness. This has the side effect that setInt(`random_seed`, ...) has a direct effect on some game's behaviour where they wouldn't be fully deterministic. We guarantee that games are deterministic so we should fix the emulator random seed and eventually get rid of these two sources of stochasticity.

With the environment being fully deterministic we have to revisit ale::State. The current `cloneState`, `cloneSystemState` doesn't make much sense and we definitely don't need `load` / `save` with the current stack implementation.

I propose we have a single `cloneState` / `restoreState` that serializes the current state **WITH** the environment (i.e., sticky action) RNG. The ALE states that sticky actions should be thought of as part of the underlying environment and exposing a method for selectively including environment stochasticity seems out of place. If the user wants a fully deterministic environment they can disable sticky actions, e.g., in planning.